Correct entity Configuration
Change projection type to be a byte instead of smallint

+++ src/SFA.DAS.Forecasting.Data/Configuration/AccountProjection.cs
@@ -13,7 +13,7 @@ namespace SFA.DAS.Forecasting.Data.Configuration
             builder.Property(x => x.Id).HasColumnName(@"Id").HasColumnType("bigint").IsRequired().ValueGeneratedOnAdd();
             builder.Property(x => x.AccountId).HasColumnName(@"EmployerAccountId").HasColumnType("bigint").IsRequired();
             builder.Property(x => x.ProjectionCreationDate).HasColumnName(@"ProjectionCreationDate").HasColumnType("datetime").IsRequired();
-            builder.Property(x => x.ProjectionGenerationType).HasColumnName(@"ProjectionGenerationType").HasColumnType("smallint").IsRequired();
+            builder.Property(x => x.ProjectionGenerationType).HasColumnName(@"ProjectionGenerationType").HasColumnType("byte").IsRequired();
             builder.Property(x => x.Month).HasColumnName(@"Month").HasColumnType("smallint").IsRequired();
             builder.Property(x => x.Year).HasColumnName(@"Year").HasColumnType("int").IsRequired();
